GH-1172: extract vscode/ into internal sub-package
Move vscode.go and vscode_test.go into pkg/orchestrator/internal/vscode/.
The only external dependency (logf) is passed as a Logger function parameter.
Root package retains thin VscodePush/VscodePop delegation methods and a
VscodeManager interface at the boundary.
